organise test running and test registration
