+++ test-harness/py_harness/test_cases/base.py
@@ -2,15 +2,17 @@ import logging
 import time
 from typing import List
 from host import LinkedCloneHost
-from . import TestCase
+from .test_case import TestCase, registered_test_cases
 
 
 class BaseTestCase(TestCase):
 
-    def __init__(self, test_case_name: str, linked_clone_hosts: List[LinkedCloneHost]):
-        super().__init__(test_case_name, linked_clone_hosts)
+    def __init__(self, linked_clone_hosts: List[LinkedCloneHost]):
+        super().__init__("base", linked_clone_hosts)
 
     def test_case(self) -> bool:
         logging.info("sleeping momentarily to simulate tests ...")
         time.sleep(5)
         return True
+
+registered_test_cases.append(BaseTestCase)
